feat: enhance utility functions for market analysis and refactor structural noise analysis
